chore: Refactor data_handler
