+++ src/data_server/mod.rs
@@ -1,4 +1,5 @@
 mod auth;
+pub mod data_handler;
 pub mod s3service;
 pub mod server;
 pub mod utils;
